+++ mslib/msui/multiple_flightpath_dockwidget.py
@@ -25,12 +25,9 @@
     limitations under the License.
 """
 import random
-
 import requests
-import functools
 import json
 from PyQt5 import QtWidgets, QtGui, QtCore
-from PyQt5.QtWidgets import QVBoxLayout, QGridLayout
 from PyQt5.QtCore import Qt
 from mslib.msui.qt5 import ui_multiple_flightpath_dockwidget as ui
 from mslib.msui import flighttrack as ft
@@ -39,6 +36,7 @@ from mslib.utils.verify_user_token import verify_user_token
 from mslib.utils.qt import Worker
 from mslib.utils.config import config_loader
 from urllib.parse import urljoin
+from mslib.utils.colordialog import CustomColorDialog
 
 
 class QMscolabOperationsListWidgetItem(QtWidgets.QListWidgetItem):
@@ -112,58 +110,6 @@ class MultipleFlightpath:
                 elem.remove()
         self.patches = []
         self.map.ax.figure.canvas.draw()
-
-
-class CustomColorDialog(QtWidgets.QDialog):
-    """
-    Custom Color Dialog to select color from predefined swatches.
-    """
-    color_selected = QtCore.pyqtSignal(QtGui.QColor)
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.setWindowTitle("Select Color")
-        self.setFixedSize(350, 300)
-
-        self.colors = [
-            "#800000", "#c31f59", "#f59757", "#fde442", "#0000ff",
-            "#60c36e", "#65d8f2", "#a446be", "#f15aea", "#b9babb",
-            "#e6194B", "#d2cf94", "#356e33", "#f58231", "#2c2c2c",
-            "#000075", "#9A6324", "#808000", "#000000", "#f8cbdc"
-        ]
-
-        layout = QVBoxLayout()
-        # Color swatches layout
-        swatch_layout = QGridLayout()
-        self.color_buttons = []
-
-        for i, color in enumerate(self.colors):
-            button = QtWidgets.QPushButton()
-            button.setFixedSize(50, 50)
-            button.setStyleSheet(f"background-color: {color}")
-            button.clicked.connect(functools.partial(self.on_color_clicked, color))
-            row = i // 5
-            col = i % 5
-            swatch_layout.addWidget(button, row, col)
-
-        # Add "Pick Custom Color" button
-        self.custom_color_button = QtWidgets.QPushButton("Pick Custom Color")
-        self.custom_color_button.clicked.connect(self.on_custom_color_clicked)
-        self.custom_color_button.setFixedSize(325, 30)
-
-        layout.addLayout(swatch_layout)
-        layout.addWidget(self.custom_color_button)
-        self.setLayout(layout)
-
-    def on_color_clicked(self, color):
-        self.color_selected.emit(QtGui.QColor(color))
-        self.accept()
-
-    def on_custom_color_clicked(self):
-        color = QtWidgets.QColorDialog.getColor()
-        if color.isValid():
-            self.color_selected.emit(color)
-            self.accept()
 
 
 class MultipleFlightpathControlWidget(QtWidgets.QWidget, ui.Ui_MultipleViewWidget):
